Create solid component for the og-image
Finally using tailwind and a more reusable component instead of the ugly
object soup it was on v0.1. Still work to do to convert this svg to png
without making the server go boom.

Maybe instead of creating these images on the fly we can prerender the image
route and save costs (but don't optimize early).