add prerendering to og-images
Fix an issue on build where the request to get the image was to ourselves
but the server is not running on build, so for now we use the midjourney cdn
link. After moving to github we'll use the github raw url.

Using the external image gave us 1+ sec on build time.