add svg->png transformation
I know the png images look better and all, but adding 300+ms to each request
we'll have to see if it's woth it. I'm thinking on generating these as a separated
"pre:build" step and move them to the public folder so they are copied automatically
by the astro builder. Add a hash or something to avoid generating them on each build.