add real sizing for typography module
This simplifies the previous version (that didn't work) and
also introduces new complexity, especially on the css side.

Regardlees this is easier, more customizable and smaller than
the other versions. I admit the UnoCSS solution is "simpler"
but it's also less dynamic, I fell this is better for all
screen sizes, althought smaller sizes can be improved