EXPERIMENTAL: dist-chunk
So far this has been working great, with the small caveat that ALL the chunks have the `*,*::before,*::after` config.
I want to try this on vercel and see if that's a big problem or not. If not, we can proceed; if it's a problem
I will investigate on the UnoCSS community about it. Either way we achieved what we wanted: reduce CSS size.